Disable savegame tabcontrol while entering savegame description.
Fixes #687. The problem described there occured because it was
possible to request the savegame description dialog again by clicking
"Save" instead of pressing "Enter". In addition, it turned out it is
possible to bury the savegame description dialog by clicking on the
savegame menu in the background again. This commit disables the
savegame menu while the savegame description dialog is open, thereby
preventing both problems.