Compile in CEGUI statically by default.
CEGUI’s development appears to have halted. The last stable version is
0.8.7 from 2016. I tried to compile TSC with the latest commit on
CEGUI’s master branch, but there have been breaking changes so that
did not work out. However, CEGUI version 0.8.7 does not compile
unmodified under modern compilers anymore.

This commit includes the patch cegui-cpp11.patch to fix that.
CEGUI’s build file itself requests C++03 as the C++ standard to use,
but this clashes with newer versions of GCC which do not accept that
standard anymore when including <cstdint>; it results in this error:

  14.2.1/bits/c++0x_warning.h:32:2: error: #error This file requires compiler and library support for the ISO C++ 2011 standard. This support must be enabled with the -std=c++11 or -std=gnu++11 compiler options.

cegui-cpp11.patch simply changes the requested C++ standard of CEGUI
to C++11. That causes a number of warnings to be emitted, but CEGUI
does compile anyway.

The second patch, cegui-fix-factorymodule.patch, fixes this problem:

  http://cegui.org/forum/viewtopic.php?t=6473

It is a bug in CEGUI when compiled statically.

This commit makes TSC build CEGUI 0.8.7 itself and link it in
statically. This should, hopefully, for the foreseeable future resolve
problems with this aged piece of code; at least, as long as CEGUI
0.8.7 can be compiled mostly unmodified with modern compilers.

It is still possible to use a system CEGUI library. This commit
introduces the new compilation option USE_SYSTEM_CEGUI, which defaults
to OFF, but setting it to ON will return to the build system state
from before this commit.